refactor: move receipt related fixtures into test_vectors
